pyproject.toml
[tool.poetry]
name = "ps-action-manager"
version = "0.1.0"
description = "A wrapper for simplifying Photoshop Action Manager code written in Python."
authors = ["MrTeferi <freethoughtleft@gmail.com>"]
license = "MPL 2.0"
readme = "README.md"
packages = [{include = "ps_action_manager"}]

[tool.poetry.dependencies]
python = "^3.9"
photoshop-python-api = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
